requirements.txt
boto3==1.17.52
botocore==1.20.52
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Assets==2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
s3transfer==0.3.7
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.4
webassets==2.0
Werkzeug==1.0.1
